utils/requirements.txt
geoip2==4.4.0
requests==2.26.0
PyYAML==6.0
cv2box
beautifulsoup4
vmess
